stb_image模块分离，编译体积预计降350K
最小程序：804K->455K
包含图像文件读写的程序体积不变
此commit后，EGE正式进入新时代！
